requirements.txt
aiohttp==3.11.12
beautifulsoup4==4.13.3
Flask==3.1.0
lxml==5.3.1
python-dotenv==1.0.1
python-telegram-bot==21.10
selenium==4.28.1
SQLAlchemy==2.0.38
webdriver-manager==4.0.1